feat(shared): move AttachmentPolicy into mercury-core (Phase 1, Android side)
The pure attachment policy (naming hygiene, kind routing, metadata and
staging-time caps, prompt assembly) now lives once in the shared KMP core;
Android consumes it through typealiases so call sites, behavior, and all
user-visible strings are unchanged. The ported common suite asserts the
rejection reasons and AttachmentTooLargeException messages verbatim — the
strings the iOS port previously hand-mirrored by line number.

Existing Android suites pass unchanged alongside the new common tests.
iOS still uses its Swift AttachmentPolicy; replacing it with a facade over
the framework is the remaining Phase 1 step.
