+++ app/src/main/java/com/unsupportedpastels/hermesandroid/share/SharePayloadPolicy.kt
@@ -1,7 +1,7 @@
 package com.unsupportedpastels.hermesandroid.share
 
 import com.unsupportedpastels.hermesandroid.app.ComposerAttachment
-import com.unsupportedpastels.hermesandroid.attachment.AttachmentAddResult
+import com.unsupportedpastels.mercury.core.attachment.AttachmentAddResult
 import com.unsupportedpastels.hermesandroid.attachment.AttachmentPolicy
 import com.unsupportedpastels.hermesandroid.attachment.checkAdd
